+++ src/AppInstallerRepositoryCore/Rest/Schema/1_0/Interface.h
@@ -43,7 +43,7 @@ namespace AppInstaller::Repository::Rest::Schema::V1_0
         Http::HttpClientHelper::HttpRequestHeaders m_requiredRestApiHeaders;
 
     private:
-        void FilterSearchResult(const SearchRequest& request, SearchResult& result) const;
+        void FilterSearchResult(const SearchRequest& request, SearchResult& result, size_t& remainingManifestRetrievals) const;
 
         std::string m_restApiUri;
         utility::string_t m_searchEndpoint;